// src/app/modules/configuration/finance/models/search-fee-type-view.ts
export interface SearchFeeTypeView {
    Name?: Array<string>;
    displayName?: Array<string>;
    schoolFacilityName?: Array<string>;
    feeTypeCategoryName?: Array<string>;
    feeAccountName?: Array<string>;
    feeTermName?: Array<string>;
    Code?: Array<string>;
    description?: Array<string>;
    SchoolFacilityId?: Array<string>;
    FeeTermId?: Array<string>;
    FeeTypeCategoryId?: Array<string>;
    FeeAccountId?: Array<string>;
    IsSameForEntireClass?: Array<number>;
    sortBy?: string;
    sortOrder?: number;
    pageNumber?: number;
    pageSize?: number;




}